+++ app/src/main/java/com/example/soundtest/Question/MyAdapter.java
@@ -4,12 +4,19 @@ import android.content.Context;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+import android.widget.EditText;
+import android.widget.ImageView;
 import android.widget.TextView;
+import android.widget.Toast;
 
 import androidx.annotation.NonNull;
 import androidx.recyclerview.widget.RecyclerView;
 
 import com.example.soundtest.R;
+import com.google.android.gms.tasks.OnCompleteListener;
+import com.google.android.gms.tasks.Task;
+import com.google.firebase.database.DatabaseReference;
+import com.google.firebase.database.FirebaseDatabase;
 
 import java.util.ArrayList;
 
@@ -17,6 +24,8 @@ public class MyAdapter extends RecyclerView.Adapter<MyAdapter.MyViewHolder> {
 
     Context context;
     ArrayList<TypeQuestionClass> questions;
+    String answer;
+    DatabaseReference databaseReference;
 
     public MyAdapter(Context c ,ArrayList<TypeQuestionClass> p){
 
@@ -31,9 +40,30 @@ public class MyAdapter extends RecyclerView.Adapter<MyAdapter.MyViewHolder> {
     }
 
     @Override
-    public void onBindViewHolder(@NonNull MyViewHolder holder, int position) {
+    public void onBindViewHolder(@NonNull final MyViewHolder holder, final int position) {
 
         holder.userQuestion.setText(questions.get(position).getTypeQuestion());
+        holder.sendAnswerButtn.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+
+                answer = holder.typeAnswerET.getText().toString();
+                String postKey = questions.get(position).getKey();
+                databaseReference =FirebaseDatabase.getInstance().getReference();
+                databaseReference.child("Question").child(postKey).child("typeAnswer").setValue(answer)
+                        .addOnCompleteListener(new OnCompleteListener<Void>() {
+                            @Override
+                            public void onComplete(@NonNull Task<Void> task) {
+                                if(task.isSuccessful()){
+                                    Toast.makeText(context, "Thank you for your answer.", Toast.LENGTH_SHORT).show();
+                                }
+                                else {
+                                    Toast.makeText(context, "Error: "+task.getException().getMessage(), Toast.LENGTH_SHORT).show();
+                                }
+                            }
+                        });
+            }
+        });
     }
 
     @Override
@@ -44,12 +74,16 @@ public class MyAdapter extends RecyclerView.Adapter<MyAdapter.MyViewHolder> {
     class MyViewHolder extends RecyclerView.ViewHolder
 {
 
+    EditText typeAnswerET;
+    private ImageView sendAnswerButtn;
     TextView userQuestion;
 
     public MyViewHolder(@NonNull View itemView) {
         super(itemView);
 
         userQuestion = itemView.findViewById(R.id.userQuestion);
+        typeAnswerET = itemView.findViewById(R.id.typeAnswerET);
+        sendAnswerButtn = itemView.findViewById(R.id.sendAnswernBttn);
     }
 }
