+++ app/src/main/java/com/example/soundtest/Question/ScholarsAnswerActivity.java
@@ -28,9 +28,6 @@ import java.util.ArrayList;
 
 public class ScholarsAnswerActivity extends AppCompatActivity {
 
-    private EditText typeAnswerET;
-    private String answer;
-    private ImageView sendAnswerButtn;
     DatabaseReference reference;
     RecyclerView recyclerView;
     ArrayList<TypeQuestionClass> list;
@@ -41,9 +38,7 @@ public class ScholarsAnswerActivity extends AppCompatActivity {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_scholars_answer);
 
-
-        typeAnswerET = findViewById(R.id.typeAnswerET);
-        sendAnswerButtn = findViewById(R.id.sendAnswernBttn);
+        
         recyclerView = findViewById(R.id.myRecycler);
         recyclerView.setLayoutManager(new LinearLayoutManager(this));
         list = new ArrayList<TypeQuestionClass>();
@@ -69,27 +64,5 @@ public class ScholarsAnswerActivity extends AppCompatActivity {
             }
         });
 
-        sendAnswerButtn .setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-
-                answer = typeAnswerET.getText().toString();
-                FirebaseDatabase.getInstance().getReference().child("Question").child("key").child("typeAnswer").setValue(answer)
-                        .addOnCompleteListener(new OnCompleteListener<Void>() {
-                            @Override
-                            public void onComplete(@NonNull Task<Void> task) {
-
-                                if(task.isSuccessful()){
-                                    Toast.makeText(ScholarsAnswerActivity.this, "Thank you for your answer.", Toast.LENGTH_SHORT).show();
-                                }
-
-                                else{
-                                    Toast.makeText(ScholarsAnswerActivity.this, "Error: "+task.getException().getMessage(), Toast.LENGTH_SHORT).show();
-                                }
-                            }
-                        });
-
-            }
-        });
     }
 }
